refactor forms

+++ resources/views/task/index.blade.php
@@ -8,31 +8,32 @@
 @endphp
 @section('content')
     <div class="container">
+        <h1>@lang('layout.tasks')</h1>
         <div class="d-flex">
             @auth
-            <p><a class="btn btn-success" href="{{ route('tasks.create') }}">@lang('create')</a></p>
+            <p><a class="btn btn-success" href="{{ route('tasks.create') }}">@lang('layout.common.buttons.create')</a></p>
             @endauth
             <div class="ml-auto">
                 {!! Form::open(['class' => 'form-inline', 'method' => 'GET']) !!}
-                {!! Form::select('filter[status_id]', $taskStatuses, null, ['placeholder' => __('status'), 'class' => 'form-control mr-2']) !!}
-                {!! Form::select('filter[created_by_id]', $users, null, ['placeholder' => __('creator'), 'class' => 'form-control mr-2']) !!}
-                {!! Form::select('filter[assigned_to_id]', $users, null, ['placeholder' => __('assignee'), 'class' => 'form-control mr-2']) !!}
-                {!! Form::submit(__('apply'), ['class' => 'btn btn-outline-primary mr-2']) !!}
+                    {!! Form::select('filter[status_id]', $taskStatuses, null, ['placeholder' => __('layout.common.task_status'), 'class' => 'form-control mr-2']) !!}
+                    {!! Form::select('filter[created_by_id]', $users, null, ['placeholder' => __('layout.task.creator'), 'class' => 'form-control mr-2']) !!}
+                    {!! Form::select('filter[assigned_to_id]', $users, null, ['placeholder' => __('layout.task.assignee'), 'class' => 'form-control mr-2']) !!}
+                    {!! Form::submit(__('layout.task.filter.apply'), ['class' => 'btn btn-outline-primary mr-2']) !!}
+                    <a href="{{ route('tasks.index') }}" class="btn btn-outline-secondary">@lang('layout.task.filter.reset')</a>
                 {!! Form::close() !!}
             </div>
         </div>
-        <div class="row justify-content-center my-2">
             <table class="table">
                 <thead>
                 <tr>
                     <th scope="col">#</th>
-                    <th scope="col">@lang('status')</th>
-                    <th scope="col">@lang('name')</th>
-                    <th scope="col">@lang('creator')</th>
-                    <th scope="col">@lang('assignee')</th>
-                    <th scope="col">@lang('created_at')</th>
+                    <th scope="col">@lang('layout.common.task_status')</th>
+                    <th scope="col">@lang('layout.common.name')</th>
+                    <th scope="col">@lang('layout.task.creator')</th>
+                    <th scope="col">@lang('layout.task.assignee')</th>
+                    <th scope="col">@lang('layout.common.created_at')</th>
                     @auth
-                    <th scope="col">@lang('actions')</th>
+                    <th scope="col">@lang('layout.common.actions')</th>
                     @endauth
                 </tr>
                 </thead>
@@ -47,14 +48,12 @@
                         <td>{{ $task->created_at }}</td>
                         <td>
                         @auth
-                            <a href="{{ route('tasks.edit', $task) }}">@lang('edit')</a>
-                            @if((string)auth()->id() === (string)$task->created_by_id)
-                            <a href="{{ route('tasks.destroy', $task) }}"
-                               class="text-danger"
-                               data-confirm="@lang('confirm')"
-                               data-method="delete"
+                            <a href="{{ route('tasks.edit', $task) }}">@lang('layout.common.edit')</a>
+                            @if(auth()->user()->isCreator($task))
+                            <a href="{{ route('tasks.destroy', $task) }}" class="text-danger"
+                               data-confirm="@lang('layout.common.confirm_destroy')" data-method="delete"
                                rel="nofollow">
-                                @lang('destroy')
+                                @lang('layout.common.destroy')
                             </a>
                             @endif
                         @endauth
@@ -63,6 +62,5 @@
                 @endforeach
                 </tbody>
             </table>
-        </div>
     </div>
 @endsection
